Refactored benchmarking
Moved each benchmark into a class that inherits from an iterationBench
class which defines a default method that the benchmarking framework
invokes to initialize the feature list and store it into a static field
at =FEList.list= before running the benchmark. The individual benchmarks
may override that method to specify different setup functionality. Then,
the benchmarks simply assume that the list has already been initialized
at =FEList.list=. This way, the creation of the list is not included in
the benchmark time.

Furthermore, this setup allows modifying the number of features in the
feature list by just passing a paramater to =invokeMain= in the scala
tests. Thus, a single benchmark can be run easily with a variable number
of features.

Finally, each benchmark is run by a scala test class that runs it with
1-100 features. These are in individual files so that each benchmark is
run in a new JVM (this partially addresses the memory leak problem
between tests), allowing the benchmarks to be run in bulk.