Fixed index tracking of new instructions
The absolute index of instructions does /not/ need to be computed as I
had thought. It is returned by VMethodEnv.getInsnIdx.

Now the indexing is working properly such that the indeces recorded in
newInsns are the indeces of the new instructions in newEnv, so the
tagging code (un)tags the instructions properly.