Lift INVOKEDYNAMIC
There are two ways of lifting INVOKEDYNAMIC, depending on whether we
lift the invoked interface or not. The case where we lift the
interface is handled, but the other is incomplete at this stage, because
our current assumption is to lift everything.

Also added new test cases and modify config files of some existing test
cases. While executing larger programs, we found and fix a bug
regarding the use of cache when expanding V[].