Fix a corner case while optimizing store
For a local variable, it is possible that a V is stored in one branch,
but a non-V is stored in the other branch. In this case, the lifting of
load instruction is broken, especially if the non-V value is of
primitive type. To handle corner cases like this, we consider if any V
value could be stored into that local variable. If yes, we lift all
store instructions of that local variable. However, this breaks our DFA
because we need to update previous store instructions once we realize V
values could be stored. Instead of precisely backtrack to those affected
store instructions, we simply rerun the whole DFA, until we can't find
any new local variable that could store conflicting runtime values.