Fix VBlock grouping
VBlock indexing of the loop header often dangles between the VBlock id
from the loop itself and the id outside of the loop (see
testNextPoisson for example). To avoid such dangling, we store all the
updates to VBlock ids, and always pick the smallest id. This solution
implies a lexical order among blocks though.